DBリフレッシュしてね！DB変更に伴う諸々を対応
