+++ app/ContentsStory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class ContentsStory extends Model
+{
+    //
+    // protected $table = 'contents_stories';
+}
