+++ public/css/story_create.css
@@ -51,6 +51,20 @@ button {
   background: #f8f4e6;
 }
 
+.content_title {
+  text-align: center;
+  margin: 20px;
+}
+
+#download {
+  margin: 20px auto;
+  padding: 5px 20px;
+  background: #001e43;
+  color: #fff;
+  font-size: 15px;
+  display: block;
+}
+
 /* モーダル */
 /* 確認画面のモーダル */
 .modal_masc {
@@ -67,7 +81,7 @@ button {
 #modal {
   position: fixed;
   width: 40%;
-  height: 470px;
+  height: 600px;
   top: 80px;
   left: 30%;
   right: 30%;
@@ -77,13 +91,40 @@ button {
 }
 
 .conf_title {
-  margin: 20px auto;
+  margin: 30px auto;
+  text-align: center;
+  font-size: 20px;
+  font-weight: bold;
+  letter-spacing: 0.2rem;
+}
+
+.conf_title2 {
+  margin: 5px auto;
   text-align: center;
   font-size: 15px;
 }
 
 .conf_contents {
   margin: 10 auto;
+}
+
+.title_container {
+  margin-top: 20px;
+}
+.cof_calm {
+  margin-top: 20px;
+  font-size: 15px;
+}
+
+input[type="text"] {
+  border: 0;
+  padding: 5px 10px;
+  font-size: 15px;
+  font-family: Arial, sans-serif;
+  color: #000;
+  border: solid 1px #ccc;
+  margin: 0 0 20px;
+  width: 300px;
 }
 
 .conf_img_container {
@@ -120,9 +161,6 @@ input[type="submit"] {
   margin: 10px;
   font-size: 15px;
   font-weight: bold;
-}
-
-.submit_btn:hover {
 }
 /* 確認画面のモーダル */
 .hidden {
